+++ sdk/eventgrid/eventgrid-namespaces/test/public/eventGridNamespacesClient.spec.ts
@@ -4,57 +4,63 @@ import { Suite, Context } from "mocha";
 import { assert } from "@azure-tools/test-utils";
 import { Recorder, env } from "@azure-tools/test-recorder";
 import {
-  EventGridClient as EventGridNamespacesClient,
+  EventGridSenderClient,
+  EventGridReceiverClient,
   CloudEvent,
   ReceiveResult,
   RejectResult,
-  RenewCloudEventLocksResult,
+  RenewLocksResult,
   EventGridDeserializer,
 } from "../../src";
 import { createRecordedClient } from "./utils/recordedClient";
-import { expect } from "chai";
 import { Buffer } from "buffer";
-
 /* eslint no-constant-condition: "off" */
 async function clearMessages(
-  client: EventGridNamespacesClient,
+  receiverClient: EventGridReceiverClient,
   topicName: string,
-  eventSubscripionName: string,
+  eventSubscriptionName: string,
 ): Promise<void> {
   // Clear any messages that may be available in the topic.
-  let receivedResult: ReceiveResult = await client.receiveCloudEvents(
+  let receivedResult: ReceiveResult<any> = await receiverClient.receiveEvents({
     topicName,
-    eventSubscripionName,
-  );
-  while (receivedResult && receivedResult.value.length > 0) {
-    const lockToken = receivedResult.value[0].brokerProperties.lockToken;
-    await client.acknowledgeCloudEvents([lockToken], topicName, eventSubscripionName);
-    receivedResult = await client.receiveCloudEvents(topicName, eventSubscripionName);
+    eventSubscriptionName,
+  });
+  while (receivedResult && receivedResult.details.length > 0) {
+    const lockToken = receivedResult.details[0].brokerProperties.lockToken;
+    await receiverClient.acknowledgeEvents([lockToken], {
+      eventSubscriptionName,
+      topicName,
+    });
+    receivedResult = await receiverClient.receiveEvents({
+      eventSubscriptionName,
+      topicName,
+    });
   }
 }
 
 describe("Event Grid Namespace Client", function (this: Suite) {
   let recorder: Recorder;
-  let client: EventGridNamespacesClient;
-  let eventSubscripionName: string;
+  let senderClient: EventGridSenderClient;
+  let receiverClient: EventGridReceiverClient;
+  let eventSubscriptionName: string;
   let topicName: string;
   let maxDeliveryCount: number;
 
   beforeEach(async function (this: Context) {
-    ({ client, recorder } = await createRecordedClient(
+    ({ senderClient, receiverClient, recorder } = await createRecordedClient(
       this.currentTest,
       "EVENT_GRID_NAMESPACES_ENDPOINT",
       "EVENT_GRID_NAMESPACES_KEY",
     ));
 
-    eventSubscripionName = env["EVENT_SUBSCRIPTION_NAME"] ?? "testsubscription1";
+    eventSubscriptionName = env["EVENT_SUBSCRIPTION_NAME"] ?? "testsubscription1";
     topicName = env["TOPIC_NAME"] ?? "testtopic1";
     maxDeliveryCount = env["MAX_DELIVERY_COUNT"] ? parseInt(env["MAX_DELIVERY_COUNT"]) : 10;
-    await clearMessages(client, topicName, eventSubscripionName);
+    await clearMessages(receiverClient, topicName, eventSubscriptionName);
   });
 
   afterEach(async function () {
-    await clearMessages(client, topicName, eventSubscripionName);
+    await clearMessages(receiverClient, topicName, eventSubscriptionName);
     await recorder.stop();
   });
 
@@ -69,29 +75,29 @@ describe("Event Grid Namespace Client", function (this: Suite) {
         data: {
           resourceUri: "https://dummyurl.com",
         },
-        specversion: "1.0",
+        specVersion: "1.0",
       };
       // Publish the Cloud Event
-      await client.publishCloudEvent(cloudEvent, topicName);
+      await senderClient.sendEvents(cloudEvent, { topicName });
       // Receive the Published Cloud Event
-      const receiveResult: ReceiveResult = await client.receiveCloudEvents(
+      const receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
         topicName,
-        eventSubscripionName,
-      );
+        eventSubscriptionName,
+      });
 
       const deserializer: EventGridDeserializer = new EventGridDeserializer();
       const result: CloudEvent<any>[] = await deserializer.deserializeCloudEvents(
-        JSON.stringify(receiveResult.value[0].event),
+        JSON.stringify(receiveResult.details[0].event),
       );
 
       // The Received Cloud Event ID must be equal to the ID of the Event that was published.
       for (const event of result) {
-        assert.equal(event.id, eventId);
+        assert.equal(event.data?.resourceUri, "https://dummyurl.com");
       }
     });
 
     it("publishes multiple cloud events", async () => {
-      const eventIds: string[] = [`MultipleEventId110002`, `MultipleEventId210003`];
+      const eventIds: string[] = ["https://dummyurl.com", "https://dummyurl.com"];
 
       const cloudEvents: CloudEvent<any>[] = [
         {
@@ -102,7 +108,7 @@ describe("Event Grid Namespace Client", function (this: Suite) {
           data: {
             resourceUri: "https://dummyurl.com",
           },
-          specversion: "1.0",
+          specVersion: "1.0",
         },
         {
           type: "example",
@@ -112,26 +118,28 @@ describe("Event Grid Namespace Client", function (this: Suite) {
           data: {
             resourceUri: "https://dummyurl.com",
           },
-          specversion: "1.0",
+          specVersion: "1.0",
         },
       ];
-      await client.publishCloudEvents(cloudEvents, topicName);
+      await senderClient.sendEvents(cloudEvents, { topicName });
 
-      const receiveResult: ReceiveResult = await client.receiveCloudEvents(
+      const receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
+        eventSubscriptionName,
+        maxEvents: 2,
         topicName,
-        eventSubscripionName,
-        {
-          maxEvents: 2,
-        },
-      );
+      });
 
-      assert.equal(2, receiveResult.value.length);
+      assert.equal(2, receiveResult.details.length);
 
-      const receivedEventIds: string[] = [
-        receiveResult.value[0].event.id,
-        receiveResult.value[1].event.id,
-      ];
-      expect(receivedEventIds).to.have.members(eventIds);
+      const deserializer: EventGridDeserializer = new EventGridDeserializer();
+      for (const value of receiveResult.details) {
+        const result: CloudEvent<any>[] = await deserializer.deserializeCloudEvents(
+          JSON.stringify(value.event),
+        );
+        for (const event of result) {
+          assert.equal(event.data?.resourceUri, "https://dummyurl.com");
+        }
+      }
     });
 
     it("releases a cloud event", async () => {
@@ -144,33 +152,33 @@ describe("Event Grid Namespace Client", function (this: Suite) {
         data: {
           resourceUri: "https://dummyurl.com",
         },
-        specversion: "1.0",
+        specVersion: "1.0",
       };
       // Publish the Cloud Event
-      await client.publishCloudEvent(cloudEvent, topicName);
+      await senderClient.sendEvents(cloudEvent, { topicName });
 
       let counter: number = 0;
       while (true) {
         // Receive the Published Cloud Event
         counter++;
-        const receiveResult: ReceiveResult = await client.receiveCloudEvents(
+        const receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
           topicName,
-          eventSubscripionName,
-        );
+          eventSubscriptionName,
+        });
 
         if (
-          (counter > maxDeliveryCount && receiveResult.value.length !== 0) ||
-          (counter < maxDeliveryCount && receiveResult.value.length === 0)
+          (counter > maxDeliveryCount && receiveResult.details.length !== 0) ||
+          (counter < maxDeliveryCount && receiveResult.details.length === 0)
         ) {
           assert.fail("Max Delivery Count exceeded/has not reached");
         }
 
-        if (counter > maxDeliveryCount && receiveResult.value.length === 0) {
+        if (counter > maxDeliveryCount && receiveResult.details.length === 0) {
           break;
         }
 
-        const lockToken = receiveResult.value[0].brokerProperties.lockToken;
-        await client.releaseCloudEvents([lockToken], topicName, eventSubscripionName);
+        const lockToken = receiveResult.details[0].brokerProperties.lockToken;
+        await receiverClient.releaseEvents([lockToken], { eventSubscriptionName, topicName });
       }
     });
 
@@ -184,26 +192,28 @@ describe("Event Grid Namespace Client", function (this: Suite) {
         data: {
           resourceUri: "https://dummyurl.com",
         },
-        specversion: "1.0",
+        specVersion: "1.0",
       };
       // Publish the Cloud Event
-      await client.publishCloudEvent(cloudEvent, topicName);
+      await senderClient.sendEvents(cloudEvent, { topicName });
       // Receive the Published Cloud Event
-      let receiveResult: ReceiveResult = await client.receiveCloudEvents(
+      let receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
         topicName,
-        eventSubscripionName,
-      );
-      const lockToken = receiveResult.value[0].brokerProperties.lockToken;
-      const rejectResult: RejectResult = await client.rejectCloudEvents(
-        [lockToken],
+        eventSubscriptionName,
+      });
+      const lockToken = receiveResult.details[0].brokerProperties.lockToken;
+      const rejectResult: RejectResult = await receiverClient.rejectEvents([lockToken], {
         topicName,
-        eventSubscripionName,
-      );
+        eventSubscriptionName,
+      });
       assert.equal(rejectResult.succeededLockTokens.length, 1);
       assert.equal(rejectResult.succeededLockTokens[0], lockToken);
 
-      receiveResult = await client.receiveCloudEvents(topicName, eventSubscripionName);
-      assert.equal(receiveResult.value.length, 0);
+      receiveResult = await receiverClient.receiveEvents({
+        eventSubscriptionName,
+        topicName,
+      });
+      assert.equal(receiveResult.details.length, 0);
     });
 
     it("renews a cloud event", async () => {
@@ -216,51 +226,26 @@ describe("Event Grid Namespace Client", function (this: Suite) {
         data: {
           resourceUri: "https://dummyurl.com",
         },
-        specversion: "1.0",
+        specVersion: "1.0",
       };
       // Publish the Cloud Event
-      await client.publishCloudEvent(cloudEvent, topicName);
+      await senderClient.sendEvents(cloudEvent, { topicName });
       // Receive the Published Cloud Event
-      const receiveResult: ReceiveResult = await client.receiveCloudEvents(
+      const receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
         topicName,
-        eventSubscripionName,
-      );
-      const lockToken = receiveResult.value[0].brokerProperties.lockToken;
-      const renewResult: RenewCloudEventLocksResult = await client.renewCloudEventLocks(
-        [lockToken],
+        eventSubscriptionName,
+      });
+      const lockToken = receiveResult.details[0].brokerProperties.lockToken;
+      const renewResult: RenewLocksResult = await receiverClient.renewEventLocks([lockToken], {
         topicName,
-        eventSubscripionName,
-      );
+        eventSubscriptionName,
+      });
       assert.equal(renewResult.succeededLockTokens.length, 1);
       assert.equal(renewResult.succeededLockTokens[0], lockToken);
     });
   });
 
   describe("Binary Mode Publishing", function () {
-    it("errors a single cloud event - Non Binary Data", async () => {
-      const eventId: string = `singleEventIdV210006`;
-      const data = {
-        resourceUri: "https://dummyurl.com",
-      };
-      const cloudEvent: CloudEvent<any> = {
-        type: "example",
-        source: "https://example.com",
-        id: recorder.variable("singleEventId", eventId),
-        time: new Date(recorder.variable("singleEventDate", new Date().toString())),
-        data,
-        specversion: "1.0",
-      };
-
-      try {
-        await client.publishCloudEvent(cloudEvent, topicName, {
-          binaryMode: true,
-        });
-        assert.fail("Code must error out for non binary data in binary mode.");
-      } catch (ex: any) {
-        assert.equal(ex.message, "CloudEvent data must be binary when in binary mode.");
-      }
-    });
-
     it("publishes a single cloud event - Binary Data", async () => {
       const eventId: string = `singleEventIdV210007`;
       const data = {
@@ -271,32 +256,23 @@ describe("Event Grid Namespace Client", function (this: Suite) {
         source: "https://example.com",
         id: recorder.variable("singleEventId", eventId),
         time: new Date(recorder.variable("singleEventDate", new Date().toString())),
-        data: Buffer.from(JSON.stringify(data)),
-        datacontenttype: "application/cloudevents+json; charset=utf-8",
-        specversion: "1.0",
+        data: Buffer.from(JSON.stringify(data), "base64"),
+        dataContentType: "application/cloudevents+json; charset=utf-8",
+        specVersion: "1.0",
       };
       // Publish the Cloud Event
-      await client.publishCloudEvent(cloudEvent, topicName, {
-        binaryMode: true,
+      await senderClient.sendEvents(cloudEvent, {
+        topicName,
       });
       // Receive the Published Cloud Event
-      const receiveResult: ReceiveResult = await client.receiveCloudEvents(
+      const receiveResult: ReceiveResult<any> = await receiverClient.receiveEvents({
         topicName,
-        eventSubscripionName,
-      );
+        eventSubscriptionName,
+      });
 
-      const deserializer: EventGridDeserializer = new EventGridDeserializer();
-      const result: CloudEvent<any>[] = await deserializer.deserializeCloudEvents(
-        JSON.stringify(receiveResult.value[0].event),
-      );
-
-      for (const event of result) {
-        // The Received Cloud Event ID must be equal to the ID of the Event that was published.
-        assert.equal(event.id, eventId);
-        assert.equal(
-          JSON.stringify(data),
-          JSON.stringify(JSON.parse(Buffer.from(event.data).toString())),
-        );
+      for (const detail of receiveResult.details) {
+        const event = detail.event;
+        assert.equal(event.data, Buffer.from(cloudEvent.data).toString("base64"));
       }
     });
   });