+++ sdk/monitor/monitor-opentelemetry-exporter/CHANGELOG.md
@@ -7,6 +7,10 @@
 - Fix issue with `disableOfflineStorage` being set sending error messages to customer apps.
 - Fix issue with `disableOfflineStorage` not applying to Statsbeat senders.
 
+### Other Changes
+
+- Client generated Span event exceptions no longer generate trace telemetry.
+
 ## 1.0.0-beta.23 (2024-05-10)
 
 ### Features Added